+++ src/gospatial/main/gospatial/main.go
@@ -55,6 +55,7 @@ func main() {
 			app.Info.Println("Waiting for sockets to close...")
 			for {
 				if len(app.Hub.Sockets) == 0 {
+					app.DB.Backup()
 					app.Info.Println("Shutting down...")
 					os.Exit(0)
 				}
@@ -65,6 +66,7 @@ func main() {
 	// Initiate Database
 	app.DB = app.Database{File: "./" + database + ".db"}
 	app.DB.Init()
+	app.DB.Backup()
 
 	router := app.NewRouter()