+++ test_runner.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+echo "Setting workspace"
+export GOPATH="`pwd`"
+
+cd src/gospatial/app
+go test -bench=. -test.benchmem