+++ static/find/L.Find.Draw.js
@@ -173,6 +173,15 @@ L.Find.Draw = L.Class.extend({
 		var feature = this.drawnItems._layers[id];
 		var payload = feature.toGeoJSON();
 		payload.properties = this.getProperties();
+		// add date_created & date_modified to feature properties
+		var now = Date();
+		if (!payload.properties.hasOwnProperty("date_created")) {
+			payload.properties.date_created = now.toISOString();
+		}
+		if (!payload.properties.hasOwnProperty("date_modified")) {
+			payload.properties.date_modified = now.toISOString();
+		}
+		// Send request
 		results = this.postRequest(
 			'/api/v1/layer/' + $('#layers').val() + '/feature',
 			JSON.stringify(payload)